Remove all usage of garage's internal tf.space
* All algorithms and policies now correctly depend on gym.spaces

+++ garage/exploration_strategies/ou_strategy.py
@@ -59,7 +59,8 @@ class OUStrategy(ExplorationStrategy, Serializable):
 if __name__ == "__main__":
     ou = OUStrategy(
         env_spec=AttrDict(
-            action_space=gym.spaces.Box(low=-1, high=1, shape=(1, ))),
+            action_space=gym.spaces.Box(
+                low=-1, high=1, shape=(1, ), dtype=np.float32)),
         mu=0,
         theta=0.15,
         sigma=0.3,
